# Added bespoke error for null cases